+++ ChargingApp/DTOs/CurrencyDto.cs
@@ -5,5 +5,5 @@ namespace ChargingApp.DTOs;
 public class CurrencyDto
 {
    [Required]  public string Name { get; set; }
-   [Required] public double ValuePerDollar { get; set; }
+   [Required] public decimal ValuePerDollar { get; set; }
 }
