+++ ChargingApp/DTOs/DebitDto.cs
@@ -4,6 +4,6 @@ public class DebitDto
 {
     public string Username { get; set; }
     public string UserEmail { get; set; }
-    public double DebitValue { get; set; }
+    public decimal DebitValue { get; set; }
     public DateTime Date { get; set; }
 }
