+++ ChargingApp/DTOs/NewProductDto.cs
@@ -6,11 +6,11 @@ public class NewProductDto
     public string? ArabicName { get; set; }
     public int MinimumQuantityAllowed { get; set; }
     public IFormFile PhotoFile { get; set; }
-    public double Price { get; set; }
+    public decimal Price { get; set; }
 }
 
 public class NewProductWithQuantityDto
 {
-    public ICollection<double> PriceList { get; set; }
-    public ICollection<double> QuantityList { get; set; }
+    public ICollection<decimal> PriceList { get; set; }
+    public ICollection<decimal> QuantityList { get; set; }
 }
