+++ ChargingApp/DTOs/PaymentAdminDto.cs
@@ -7,7 +7,7 @@ public class PaymentAdminDto
     public int Id { get; set; }
     public DateTime CreatedDate { get; set; } 
     public string? Notes{ get; set; }
-    public double AddedValue { get; set; }
+    public decimal AddedValue { get; set; }
     public string Photo { get; set; }
     public string? Username { get; set; }
     public string PaymentType { get; set; }// office or company or usdt
