+++ ChargingApp/DTOs/PendingOrderDto.cs
@@ -3,10 +3,10 @@
 public class PendingOrderDto
 {
     public int Id { get; set; }
-    public double TotalPrice { get; set; }
+    public decimal TotalPrice { get; set; }
     public string PlayerId { get; set; }
     public DateTime CreatedAt { get; set; }
-    public double Quantity { get; set; } = 1;
+    public decimal Quantity { get; set; } = 1;
     public string ProductEnglishName { get; set; }
     public string ProductArabicName { get; set; }
     public string UserName { get; set; }
