+++ ChargingApp/Entity/Currency.cs
@@ -4,5 +4,5 @@ public class Currency
 {
     public int Id { get; set; }
     public string Name { get; set; }
-    public double ValuePerDollar { get; set; }
+    public decimal ValuePerDollar { get; set; }
 }
