+++ ChargingApp/Interfaces/IApiService.cs
@@ -6,7 +6,7 @@ public interface IApiService
 {
     public Task<bool> CheckProductByIdIfExistAsync(int id, string baseUrl, string token);
 
-    public Task<(bool Success, string Message, int OrderId)> SendOrderAsync(int productId, double qty, string playerId,
+    public Task<(bool Success, string Message, int OrderId)> SendOrderAsync(int productId, decimal qty, string playerId,
         string baseUrl, string token);
 
     public Task<(bool Succeed, string Status )> CheckOrderStatusAsync(int orderId, string baseUrl, string token);
