+++ ChargingApp/Interfaces/ISpecificPriceForUserRepository.cs
@@ -4,7 +4,7 @@ namespace ChargingApp.Interfaces;
 
 public interface ISpecificPriceForUserRepository
 {
-    public Task<double?> GetProductPriceForUserAsync(int productId, AppUser user);
+    public Task<decimal?> GetProductPriceForUserAsync(int productId, AppUser user);
     public void AddProductPriceForUser(SpecificPriceForUser newRow);
     public void UpdateProductPriceForUser(SpecificPriceForUser newRow);
     public Task<SpecificPriceForUser?> GetPriceForUserAsync(string email , int vipLevel, int productId);
